Small formatting and layout  changes
