+++ MobileApp/city_problems/lib/main.dart
@@ -7,8 +7,8 @@ import 'package:city_problems/firebase_options.dart';
 import 'package:city_problems/models/index.dart';
 import 'package:city_problems/presentation/home.dart';
 import 'package:city_problems/presentation/home_page.dart';
-import 'package:city_problems/presentation/login.dart';
-import 'package:city_problems/presentation/signup.dart';
+import 'package:city_problems/presentation/login_page.dart';
+import 'package:city_problems/presentation/signup_page.dart';
 import 'package:city_problems/reducer/reducer.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
