+++ MobileApp/city_problems/lib/presentation/signup_page.dart
@@ -5,20 +5,21 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 
-class LoginPage extends StatefulWidget {
-  const LoginPage({super.key});
+class SignupPage extends StatefulWidget {
+  const SignupPage({super.key});
 
   @override
-  State<LoginPage> createState() => _LoginPageState();
+  State<SignupPage> createState() => _SignupPageState();
 }
 
-class _LoginPageState extends State<LoginPage> {
+class _SignupPageState extends State<SignupPage> {
   @override
   Widget build(BuildContext context) {
     final TextEditingController emailController = TextEditingController();
     final TextEditingController passwordController = TextEditingController();
+    final TextEditingController nameController = TextEditingController();
     return UserContainer(
-      builder: (BuildContext context, AppUser? appUser) {
+      builder: (BuildContext context, AppUser? user) {
         return Scaffold(
           resizeToAvoidBottomInset: false,
           body: Padding(
@@ -27,7 +28,7 @@ class _LoginPageState extends State<LoginPage> {
               child: Column(
                 children: <Widget>[
                   const Text(
-                    'Login',
+                    'Sign Up',
                     style: TextStyle(
                       color: Color.fromRGBO(80, 151, 255, 1),
                       fontSize: 60,
@@ -62,7 +63,7 @@ class _LoginPageState extends State<LoginPage> {
                     ),
                   ),
                   const SizedBox(
-                    height: 50,
+                    height: 20,
                   ),
                   Material(
                     elevation: 5,
@@ -89,7 +90,34 @@ class _LoginPageState extends State<LoginPage> {
                     ),
                   ),
                   const SizedBox(
-                    height: 30,
+                    height: 20,
+                  ),
+                  Material(
+                    elevation: 5,
+                    shadowColor: const Color.fromRGBO(80, 151, 255, 1),
+                    borderRadius: BorderRadius.circular(25),
+                    child: TextField(
+                      controller: nameController,
+                      keyboardType: TextInputType.text,
+                      obscureText: true,
+                      decoration: InputDecoration(
+                        border: OutlineInputBorder(
+                          borderRadius: BorderRadius.circular(25),
+                          borderSide: BorderSide.none,
+                        ),
+                        filled: true,
+                        hintText: 'name',
+                        hintStyle: const TextStyle(
+                          color: Color.fromRGBO(132, 150, 155, 1),
+                          fontFamily: 'ebrima',
+                          fontWeight: FontWeight.bold,
+                        ),
+                        fillColor: Colors.white,
+                      ),
+                    ),
+                  ),
+                  const SizedBox(
+                    height: 15,
                   ),
                   Row(
                     mainAxisAlignment: MainAxisAlignment.center,
@@ -103,16 +131,16 @@ class _LoginPageState extends State<LoginPage> {
                             fontSize: 14,
                           ),
                           children: <TextSpan>[
-                            const TextSpan(text: "Haven't account?  "),
+                            const TextSpan(text: 'Have an account?  '),
                             TextSpan(
-                              text: 'Sign-up',
+                              text: 'Login',
                               style: const TextStyle(
                                 color: Color.fromRGBO(50, 135, 237, 1),
                                 fontSize: 16,
                               ),
                               recognizer: TapGestureRecognizer()
                                 ..onTap = () {
-                                  Navigator.pushNamed(context, '/signup');
+                                  Navigator.pushNamed(context, '/login');
                                 },
                             ),
                           ],
@@ -126,11 +154,13 @@ class _LoginPageState extends State<LoginPage> {
                   Builder(
                     builder: (BuildContext context) {
                       return ElevatedButton(
-                        onPressed: () async{
-                          final Login action = Login(
+                        onPressed: () async {
+                          final SignUp action = SignUp(
                             email: emailController.text,
                             password: passwordController.text,
+                            displayName: nameController.text,
                           );
+
                           StoreProvider.of<AppState>(context).dispatch(action);
                         },
                         style: ElevatedButton.styleFrom(
@@ -141,7 +171,7 @@ class _LoginPageState extends State<LoginPage> {
                           height: 60,
                           child: Center(
                             child: Text(
-                              'LOGIN',
+                              'SIGN UP',
                               style: TextStyle(
                                 color: Colors.white,
                                 fontSize: 35,
