+++ MobileApp/city_problems/lib/reducer/auth_reducer.dart
@@ -5,7 +5,7 @@ import 'package:redux/redux.dart';
 Reducer<AuthState> authReducer = combineReducers(<Reducer<AuthState>>[
   TypedReducer<AuthState,LoginSuccessful>(_loginSuccessful),
   TypedReducer<AuthState,SignUpSuccessful>(_signUpSuccessful),
-  TypedReducer<AuthState, InitializeUserSuccessful>(_initializeUserSuccessful),
+  TypedReducer<AuthState,InitializeUserSuccessful>(_initializeUserSuccessful),
 ]);
 AuthState _loginSuccessful(AuthState state, LoginSuccessful action) {
   return state.copyWith(
